+++ server/src/dtsTypes/bindings/zephyr/loader.ts
@@ -395,9 +395,65 @@ export class ZephyrBindingsLoader {
 			const base = bindings.find(
 				(b) => basename(b.filePath) === `base.yaml`,
 			);
-			const baseType = base
-				? convertBindingToType(base, node)
-				: undefined;
+			let baseType: NodeType | undefined;
+			if (base && node.pathString === '/zephyr,user') {
+				const baseCopy = {
+					...base,
+					properties: { ...base.properties },
+				};
+
+				node.properties.forEach((p) => {
+					if (p.name.endsWith('-gpios')) {
+						baseCopy.properties[p.name] = {
+							name: p.name,
+							type: 'phandle-array',
+						};
+					} else if (p.name === 'handle') {
+						baseCopy.properties[p.name] = {
+							name: p.name,
+							type: 'phandle',
+						};
+					} else if (p.name === 'handles') {
+						baseCopy.properties[p.name] = {
+							name: p.name,
+							type: 'phandles',
+						};
+					} else if (p.name === 'boolean') {
+						baseCopy.properties[p.name] = {
+							name: p.name,
+							type: 'boolean',
+						};
+					} else if (p.name === 'bytes') {
+						baseCopy.properties[p.name] = {
+							name: p.name,
+							type: 'uint8-array',
+						};
+					} else if (p.name === 'number') {
+						baseCopy.properties[p.name] = {
+							name: p.name,
+							type: 'int',
+						};
+					} else if (p.name === 'numbers') {
+						baseCopy.properties[p.name] = {
+							name: p.name,
+							type: 'array',
+						};
+					} else if (p.name === 'string') {
+						baseCopy.properties[p.name] = {
+							name: p.name,
+							type: 'string',
+						};
+					} else if (p.name === 'strings') {
+						baseCopy.properties[p.name] = {
+							name: p.name,
+							type: 'string-array',
+						};
+					}
+				});
+				baseType = convertBindingToType(baseCopy, node);
+			}
+
+			baseType ??= base ? convertBindingToType(base, node) : undefined;
 			if (baseType) {
 				baseType.warnMismatchProperties = false;
 				const compat = baseType.properties.find(